# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: ANAS Web API
  version: 0.8.0-m5
  description: |
    M2 API for the ANAS administration daemon. This contract is versioned
    independently from the CLI JSON contract: HTTP responses use
    `anas.dev/api/v1`, while the CLI continues to use `anas.dev/cli/v1`.

    Workspaces are registered when `anasd` starts. Callers address them only by
    opaque workspace ID; request and response DTOs never contain host filesystem
    paths. M1A adds bootstrap exchange and local-owner authentication while the
    workspace surface remains read-only. Route visibility and transport are
    state-dependent. M1B adds durable job history and event replay, the
    certificate-bound bootstrap-to-enrollment handoff and initial local-owner
    creation, synchronous managed-configuration read/validation/atomic save,
    audited typed deployment planning, bootstrap first-apply jobs, and local
    owner step-up for full-state direct apply. M1C adds a separately built Vue
    console shell plus an independent recovery bundle, both served from
    immutable assets embedded in `anasd`. M1.5 adds a
    separate mTLS trusted-proxy listener whose verified OIDC directory identity
    receives the same management capabilities while local-account routes remain
    direct-listener only. M2 adds live container health, dependency-expanded
    Module lifecycle previews and durable start/stop/restart/rollback jobs.
...
    Saving desired configuration does not apply it. A disallowed
...
servers:
...
security: []
...
paths:
  /api/v1/workspaces/{ws}/host/actions/incus.status:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
    post:
      operationId: invokeHostIncusPreflight
      summary: Enqueue the installed read-only Incus host preflight
      description: |
        Optional surface, absent unless host_actions is explicitly enabled and
        the non-root daemon's shared host execution service started. Accepts
        only an empty JSON object (maximum 4096 bytes), never a command, path,
        credential, peer identity or installation parameter. Owner TLS, the
        normal session CSRF/Origin checks and a registered workspace are required
        on both direct and trusted-proxy listeners. A 202 response is durable
        queue admission, not successful execution or compute readiness.
        Idempotency keys are scoped to incus.status, not actor or transport;
        identical in-flight requests coalesce. Conflicting workspace or frozen
        release bindings return 409 without disclosing another job identifier.
        Execution outlives this HTTP request. Read completion at the existing
        job endpoint. Only queued jobs may be explicitly cancelled; running
        preflight has no cooperative cancellation and returns 409 to cancel.
        No package installation, configuration, uninstallation or ingress action
        is exposed by this route.
      tags: [Jobs]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/IdempotencyKey"
      security:
        - localSessionCookie: []
          csrfHeader: []
        - proxySessionCookie: []
          csrfHeader: []
      x-anas-permission: host.preflight
      x-anas-scope: workspace
      x-anas-listeners: [direct, trusted_proxy]
      x-anas-state-access:
        full: {authentication: owner, transports: [tls]}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              maxProperties: 0
      responses:
        "202":
          description: Durable new or reused job; this is not a completion receipt.
          headers:
            Location:
              required: true
              schema:
                type: string
                pattern: '^/api/v1/jobs/[A-Za-z0-9._-]+$'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequestProblem"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "405":
          $ref: "#/components/responses/PostMethodNotAllowedProblem"
        "408":
          $ref: "#/components/responses/RequestCanceledProblem"
        "409":
          $ref: "#/components/responses/ConflictProblem"
        "413":
          $ref: "#/components/responses/PayloadTooLargeProblem"
        "415":
          $ref: "#/components/responses/UnsupportedMediaTypeProblem"
        "429":
          $ref: "#/components/responses/TooManyRequestsProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        "503":
          $ref: "#/components/responses/ServiceUnavailableProblem"
        "504":
          $ref: "#/components/responses/DeadlineProblem"
...
  /api/v1/workspaces/{ws}/host/actions/confirm:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
    post:
      operationId: issueHostActionConfirmation
      summary: Issue a five-minute one-time host action confirmation
      tags: [Jobs]
      parameters:
        - $ref: "#/components/parameters/Origin"
      security:
        - localSessionCookie: []
          csrfHeader: []
        - proxySessionCookie: []
          csrfHeader: []
      x-anas-permission: host.preflight
      x-anas-scope: workspace
      x-anas-listeners: [direct, trusted_proxy]
      x-anas-state-access:
        full: {authentication: owner, transports: [tls]}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_job_id, action]
              additionalProperties: false
              properties:
                plan_job_id: {type: string}
                action: {type: string, enum: [incus.install, incus.configure, incus.enroll, incus.uninstall, incus.image-prune]}
      responses:
        "201":
          description: Confirmation token returned once; plaintext is not persisted.
          content:
            application/json:
              schema:
                type: object
                required: [api_version, token, binding_digest, expires_at]
                properties:
                  api_version: {type: string}
                  token: {type: string}
                  binding_digest: {type: string, pattern: "^[0-9a-f]{64}$"}
                  expires_at: {type: string, format: date-time}
        "400": {$ref: "#/components/responses/BadRequestProblem"}
        "401": {$ref: "#/components/responses/UnauthorizedProblem"}
        "403": {$ref: "#/components/responses/ForbiddenProblem"}
        "404": {$ref: "#/components/responses/NotFoundProblem"}
        "405": {$ref: "#/components/responses/PostMethodNotAllowedProblem"}
        "408": {$ref: "#/components/responses/RequestCanceledProblem"}
        "409": {$ref: "#/components/responses/ConflictProblem"}
        "413": {$ref: "#/components/responses/PayloadTooLargeProblem"}
        "415": {$ref: "#/components/responses/UnsupportedMediaTypeProblem"}
        "429": {$ref: "#/components/responses/TooManyRequestsProblem"}
        "500": {$ref: "#/components/responses/InternalProblem"}
        "503": {$ref: "#/components/responses/ServiceUnavailableProblem"}
        "504": {$ref: "#/components/responses/DeadlineProblem"}
...
  /api/v1/jobs:
    get:
      operationId: listJobs
      summary: List durable jobs visible to the current principal
      description: |
        Returns newest jobs first. Bootstrap and enrollment recovery sessions
        see only jobs created for the same transaction. A full owner sees jobs
        for every workspace still registered by the server. Filtering happens
        before pagination, and the cursor is an opaque stable continuation.
      tags: [Jobs]
      security:
        - bootstrapSessionCookie: []
        - localSessionCookie: []
      x-anas-permission: job.list
      x-anas-scope: workspace
      x-anas-listeners: [direct, trusted_proxy]
      x-anas-state-access:
        bootstrap: {authentication: bootstrap, transports: [plaintext, tls]}
        enrollment: {authentication: bootstrap, transports: [plaintext, tls]}
        full: {authentication: owner, transports: [tls]}
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: One authorization-filtered page of durable job summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
        "400":
          $ref: "#/components/responses/BadRequestProblem"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "405":
          $ref: "#/components/responses/GetMethodNotAllowedProblem"
        "408":
          $ref: "#/components/responses/RequestCanceledProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        "503":
          $ref: "#/components/responses/ServiceUnavailableProblem"
        "504":
          $ref: "#/components/responses/DeadlineProblem"
  /api/v1/jobs/{id}:
    parameters:
      - $ref: "#/components/parameters/JobID"
    get:
      operationId: getJob
      summary: Read a durable job and its safe final outcome
      description: |
        Applies the same workspace and principal scope as the list endpoint.
        The transport projection deliberately omits the persisted request and
        raw creator identity while retaining safe warnings, result, and error.
      tags: [Jobs]
      security:
        - bootstrapSessionCookie: []
        - localSessionCookie: []
      x-anas-permission: job.read
      x-anas-scope: workspace
      x-anas-listeners: [direct, trusted_proxy]
      x-anas-state-access:
        bootstrap: {authentication: bootstrap, transports: [plaintext, tls]}
        enrollment: {authentication: bootstrap, transports: [plaintext, tls]}
        full: {authentication: owner, transports: [tls]}
      responses:
        "200":
          description: The safe projection of one authorized durable job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequestProblem"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "405":
          $ref: "#/components/responses/GetMethodNotAllowedProblem"
        "408":
          $ref: "#/components/responses/RequestCanceledProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        "503":
          $ref: "#/components/responses/ServiceUnavailableProblem"
        "504":
          $ref: "#/components/responses/DeadlineProblem"
...
  /api/v1/workspaces/{ws}/modules/actions/{action}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - name: action
        in: path
        required: true
        schema:
          type: string
          enum: [start, stop, restart]
    post:
      operationId: manageWorkspaceModules
      summary: Preview or enqueue a dependency-safe Module lifecycle action
      description: |
        A request containing only `modules` returns the runner-expanded chain.
        To execute, repeat the exact deployment ID, digest, and affected Module
        list returned by that preview and supply Idempotency-Key. The handler
        recomputes the preview before enqueue, and the job-owned application
        service recomputes it again under the runtime lock before touching
        containers. An empty modules list means the whole active deployment.
      tags: [Modules]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/OptionalIdempotencyKey"
      security:
        - localSessionCookie: []
          csrfHeader: []
        - proxySessionCookie: []
          csrfHeader: []
      x-anas-permission: deployment.lifecycle
      x-anas-scope: workspace
      x-anas-listeners: [direct, trusted_proxy]
      x-anas-state-access:
        full: {authentication: owner, transports: [tls]}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifecycleActionRequest"
      responses:
        "200":
          description: The current dependency-expanded action preview.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecyclePreviewResponse"
        "202":
          description: The lifecycle job was durably queued, or an identical existing job was returned.
          headers:
            Location:
              description: Relative URL of the durable job resource.
              required: true
              schema:
                type: string
                pattern: "^/api/v1/jobs/[^/]+$"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentApplyResponse"
        "400":
          $ref: "#/components/responses/BadRequestProblem"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "405":
          $ref: "#/components/responses/PostMethodNotAllowedProblem"
        "408":
          $ref: "#/components/responses/RequestCanceledProblem"
        "409":
          $ref: "#/components/responses/ConflictProblem"
        "413":
          $ref: "#/components/responses/PayloadTooLargeProblem"
        "415":
          $ref: "#/components/responses/UnsupportedMediaTypeProblem"
        "428":
          $ref: "#/components/responses/PreconditionRequiredProblem"
        "429":
          $ref: "#/components/responses/TooManyRequestsProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        "503":
          $ref: "#/components/responses/ServiceUnavailableProblem"
        "504":
          $ref: "#/components/responses/DeadlineProblem"
...
  /api/v1/workspaces/{ws}/modules/{module}/actions/{action}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/ModuleName"
      - name: action
        in: path
        required: true
        schema:
          type: string
          enum: [enable, disable]
    post:
      operationId: configureWorkspaceModule
      summary: Enqueue a CAS-protected Module enable or disable action
      description: |
        The request body must be an empty JSON object. `If-Match` must contain
        the strong ETag returned by the Module list. The durable job performs a
        second CAS check and writes fail-closed audit immediately before the
        existing configuration WAL publishes the new desired state.
      tags: [Modules]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/ModuleConfigIfMatch"
      security:
        - localSessionCookie: []
          csrfHeader: []
        - proxySessionCookie: []
          csrfHeader: []
      x-anas-permission: module.config
      x-anas-scope: workspace
      x-anas-listeners: [direct, trusted_proxy]
      x-anas-state-access:
        full: {authentication: owner, transports: [tls]}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "202":
          description: The Module configuration job was durably queued, or an identical existing job was returned.
          headers:
            Location:
              description: Relative URL of the durable job resource.
              required: true
              schema:
                type: string
                pattern: "^/api/v1/jobs/[^/]+$"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentApplyResponse"
        "400":
          $ref: "#/components/responses/BadRequestProblem"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "405":
          $ref: "#/components/responses/PostMethodNotAllowedProblem"
        "408":
          $ref: "#/components/responses/RequestCanceledProblem"
        "409":
          $ref: "#/components/responses/ConflictProblem"
        "412":
          $ref: "#/components/responses/PreconditionProblem"
        "413":
          $ref: "#/components/responses/PayloadTooLargeProblem"
        "415":
          $ref: "#/components/responses/UnsupportedMediaTypeProblem"
        "428":
          $ref: "#/components/responses/PreconditionRequiredProblem"
        "429":
          $ref: "#/components/responses/TooManyRequestsProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        "503":
          $ref: "#/components/responses/ServiceUnavailableProblem"
        "504":
          $ref: "#/components/responses/DeadlineProblem"
...
  /api/v1/workspaces/{ws}/snapshots/{id}/actions/{action}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/SnapshotID"
      - $ref: "#/components/parameters/SnapshotAction"
    post:
      operationId: changeSnapshot
      summary: Queue snapshot pin, unpin or verification
      tags: [Maintenance]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/IdempotencyKey"
      security:
        - localSessionCookie: []
          csrfHeader: []
        - proxySessionCookie: []
          csrfHeader: []
      x-anas-permission: snapshot.change
      x-anas-scope: workspace
      x-anas-listeners: [direct, trusted_proxy]
      x-anas-state-access:
        full: {authentication: owner, transports: [tls]}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnapshotActionRequest"
      responses:
        "202": { $ref: "#/components/responses/AcceptedJob" }
        "400": { $ref: "#/components/responses/BadRequestProblem" }
        "401": { $ref: "#/components/responses/UnauthorizedProblem" }
        "403": { $ref: "#/components/responses/ForbiddenProblem" }
        "404": { $ref: "#/components/responses/NotFoundProblem" }
        "405": { $ref: "#/components/responses/PostMethodNotAllowedProblem" }
        "408": { $ref: "#/components/responses/RequestCanceledProblem" }
        "409": { $ref: "#/components/responses/ConflictProblem" }
        "413": { $ref: "#/components/responses/PayloadTooLargeProblem" }
        "415": { $ref: "#/components/responses/UnsupportedMediaTypeProblem" }
        "429": { $ref: "#/components/responses/TooManyRequestsProblem" }
        "500": { $ref: "#/components/responses/InternalProblem" }
        "503": { $ref: "#/components/responses/ServiceUnavailableProblem" }
        "504": { $ref: "#/components/responses/DeadlineProblem" }
components:
  securitySchemes:
    preAuthCSRFCookie:
      type: apiKey
      in: cookie
      name: anas_preauth_csrf
      description: |
        Host-only double-submit cookie issued by `GET /api/v1/auth/csrf`.
        Its value must exactly match `X-CSRF-Token`.
      x-anas-cookie-attributes:
        host_only: true
        http_only: true
        same_site: Strict
        path: /
        secure: state-dependent
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: The matching pre-authentication/session CSRF value; enrollment reads it from the dedicated same-origin CSRF cookie rather than a JSON exchange response.
    bootstrapSessionCookie:
      type: apiKey
      in: cookie
      name: anas_bootstrap_session
      description: |
        Host-only bootstrap session credential issued by token exchange. It is
        HttpOnly and SameSite=Strict, has Path=/ and no Domain, and is Secure
        when issued in enrollment. A promoted session authorizes only the
        enrollment handoff route for its original transaction.
      x-anas-cookie-attributes:
        host_only: true
        http_only: true
        same_site: Strict
        path: /
        secure: state-dependent
    enrollmentSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-anas_enrollment_session
      description: |
        Host-only enrollment credential issued only at the validated TLS target.
        It is Secure, HttpOnly and SameSite=Strict, has Path=/ and no Domain.
...
  parameters:
...
    WorkspaceID:
      name: ws
      in: path
      required: true
      description: Opaque ID of a workspace registered when the daemon started; never a filesystem path.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
...